checkpoint, removing more tests from old framework, deleting covid odes, moving logging -> utils submodule, creating utils submodule for common utilities and removing some old utils functions from old framework
