+++ src/dynode/__init__.py
@@ -8,7 +8,7 @@ DynODE is currently under active development and will be substantially
 refactored in the near future!
 """
 
-from . import config, infer, logging, simulate, typing
+from . import config, infer, simulate, typing, utils
 
 # Defines all the different modules able to be imported from src
-__all__ = ["config", "infer", "logging", "simulate", "typing"]
+__all__ = ["config", "infer", "utils", "simulate", "typing"]
