+++ src/dynode/config/dimension.py
@@ -14,8 +14,7 @@ from pydantic import (
 )
 from typing_extensions import Self
 
-from dynode.typing import DynodeName
-from dynode.typing.typing import UnitIntervalFloat
+from dynode.typing import DynodeName, UnitIntervalFloat
 
 from .bins import Bin, DiscretizedPositiveIntBin, WaneBin
 from .strains import Strain
