+++ src/dynode/infer/sample.py
@@ -187,80 +187,3 @@ def sample_then_resolve(
         parameters, root_params=dict(parameters)
     )
     return parameters
-
-
-def identify_distribution_indexes(
-    parameters: dict[str, Any],
-) -> dict[str, dict[str, str | tuple | None]]:
-    """Identify the locations and site names of numpyro samples.
-
-    The inverse of `sample_if_distribution()`, identifies which parameters
-    are numpyro distributions and returns a mapping between the sample site
-    names and its actual parameter name and index.
-
-    Parameters
-    ----------
-    parameters : dict[str, Any]
-        A dictionary containing keys of different parameter
-        names and values of any type.
-
-    Returns
-    -------
-    dict[str, dict[str, str | tuple[int] | None]]
-        A dictionary mapping the sample name to the dict key within `parameters`.
-        If the sampled parameter is within a larger list, returns a tuple of indexes as well,
-        otherwise None.
-
-        - key: `str`
-            Sampled parameter name as produced by `sample_if_distribution()`.
-        - value: `dict[str, str | tuple | None]`
-            "sample_name" maps to key within `parameters` and "sample_idx" provides
-            the indexes of the distribution if it is found in a list, otherwise None.
-
-    Examples
-    --------
-    >>> import numpyro.distributions as dist
-    >>> parameters = {"test": [0, dist.Normal(), 2], "example": dist.Normal()}
-    >>> identify_distribution_indexes(parameters)
-    {'test_1': {'sample_name': 'test', 'sample_idx': (1,)},
-    'example': {'sample_name': 'example', 'sample_idx': None}}
-    """
-
-    def get_index(indexes):
-        return tuple(indexes)
-
-    index_locations: dict[str, dict[str, str | tuple | None]] = {}
-    for key, param in parameters.items():
-        # if distribution, it does not have an index, so None
-        if issubclass(type(param), Dist.Distribution):
-            index_locations[key] = {"sample_name": key, "sample_idx": None}
-        # if list, check for distributions within and mark their indexes
-        elif isinstance(param, (np.ndarray, list)):
-            param = np.array(param)  # cast np.array so we get .shape
-            flat_param = np.ravel(param)  # Flatten the parameter array
-            # check for distributions inside of the flattened parameter list
-            if any(
-                [
-                    issubclass(type(param_lst), Dist.Distribution)
-                    for param_lst in flat_param
-                ]
-            ):
-                dim_idxs = np.unravel_index(
-                    np.arange(flat_param.size), param.shape
-                )
-                for i, param_lst in enumerate(flat_param):
-                    if issubclass(type(param_lst), Dist.Distribution):
-                        param_idxs = [dim_idx[i] for dim_idx in dim_idxs]
-                        index_locations[
-                            str(
-                                key
-                                + "_"
-                                + "_".join(
-                                    [str(dim_idx[i]) for dim_idx in dim_idxs]
-                                )
-                            )
-                        ] = {
-                            "sample_name": key,
-                            "sample_idx": get_index(param_idxs),
-                        }
-    return index_locations
