+++ src/dynode/utils/__init__.py
@@ -5,6 +5,8 @@ import logging
 from . import log
 from .custom_log_formatter import CustomLogFormatter
 from .log_decorator import log_decorator
+from .utils import *  # noqa: F403
+from .vis_utils import *  # noqa: F403
 
 # Fetching the global logger called dynode
 logger = logging.getLogger("dynode")
