+++ src/dynode/utils/vis_utils.py
@@ -12,10 +12,10 @@ from jax.random import PRNGKey
 from matplotlib.axes import Axes
 from matplotlib.colors import LinearSegmentedColormap
 
-from .infer import identify_distribution_indexes
-from .utils import (
+from . import (
     drop_keys_with_substring,
     flatten_list_parameters,
+    identify_distribution_indexes,
 )
